Fix a mistake of channel information

+++ src/mpeg_parser.h
@@ -70,7 +70,7 @@ typedef struct {
     int64_t                 dts;
     uint32_t                sampling_frequency;
     uint32_t                bitrate;
-    uint8_t                 channel;
+    uint16_t                channel;
     uint8_t                 layer;
     uint8_t                 bit_depth;
 } audio_sample_info_t;
